+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Added the `fill` command that fills the SD card of a Nitrokey Storage device
   with random data
+  - Added the `termion` dependency in version `1.5.5`
 - Added SD card usage information to the output of the `status` command for
   Storage devices