+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Unreleased
 - Bumped `anyhow` dependency to `1.0.39`
 - Bumped `libc` dependency to `0.2.90`
 - Bumped `nitrokey` dependency to `0.8.0`
+- Bumped `serde` dependency to `1.0.118`
 - Bumped `structopt` dependency to `0.3.21`