+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Unreleased
   `PATH` environment variable
 - Added the `fill` command that fills the SD card of a Nitrokey Storage device
   with random data
-  - Added the `termion` dependency in version `1.5.5`
+  - Added the `termion` dependency in version `1.5.6`
 - Added SD card usage information to the output of the `status` command for
   Storage devices
 - Renamed the `--{no-,}{numlock,capslock,scrollock}` options of the `config