+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Unreleased
 - Bumped `nitrokey` dependency to `0.8.0`
 - Bumped `serde` dependency to `1.0.118`
 - Bumped `structopt` dependency to `0.3.21`
+- Bumped `toml` dependency to `0.5.8`
 
 
 0.3.5